add extrakeys feature
